[Transactions] transações com eventBus

+++ FastPay.Application/Transactions/Dtos/ProcessTransactionResponse.cs
@@ -8,10 +8,9 @@ public sealed class ProcessTransactionResponse
     public string TransactionId { get; init; } = string.Empty;
     [JsonConverter(typeof(JsonStringEnumConverter))]
     public TransactionStatus Status { get; init; }
-    public decimal Balance { get; init; }
-    public decimal ReservedBalance { get; init; }
-    public decimal AvailableBalance { get; init; }
+    public long Balance { get; init; }
+    public long ReservedBalance { get; init; }
+    public long AvailableBalance { get; init; }
     public DateTime Timestamp { get; init; }
     public string? ErrorMessage { get; init; }
 }
-
